fix: do not rely on revert reason in chain service
With Ganache, when transaction execution reverts, the revert reason is
included in the thrown exeption. This is not the case for testnets
like rinkeby (https://github.com/ethers-io/ethers.js/issues/446).

The alternative to querying the nitro adjudicator for its state is to
use the [eth-revert-reason npm module](https://www.npmjs.com/package/eth-revert-reason).
However, that module is not reliable as we see the following error:
`You cannot use a blocknumber that has not yet happened.`

eth-revert-reason can be used more reliably if a delay is added before
ether-revert-reason is invoked